Use a namespace for stateTypes
