+++ src/pages/general/accountForms/LoginForm.tsx
@@ -1,20 +1,18 @@
 import React, {useRef, useState} from 'react';
 import TextInput from '../../../components/formFields/TextInput';
 import {connect} from 'react-redux';
-import {
-    Account,
-    ReduxState,
-    OAuth2Token,
-    Message,
-} from '../../../utilities/types';
+import stateTypes from '../../../utilities/types/stateTypes';
 import {authPostRequest} from '../../../utilities/axiosClients';
 import TextLink from '../../../components/links/TextLink';
 import ButtonLink from '../../../components/links/ButtonLink';
 import dispatcher from '../../../utilities/dispatcher';
 
 interface LoginFormProps {
-    logIn(oauth2_token: OAuth2Token, account: Account): void;
-    openMessage(message: Message): void;
+    logIn(
+        oauth2_token: stateTypes.OAuth2Token,
+        account: stateTypes.Account,
+    ): void;
+    openMessage(message: stateTypes.Message): void;
     closeAllMessages(): void;
 }
 
@@ -111,7 +109,7 @@ function LoginForm(props: LoginFormProps) {
     );
 }
 
-const mapStateToProps = (state: ReduxState) => ({});
+const mapStateToProps = (state: stateTypes.ReduxState) => ({});
 const mapDispatchToProps = (dispatch: any) => ({
     logIn: dispatcher.logIn(dispatch),
     openMessage: dispatcher.openMessage(dispatch),
