+++ src/pages/general/accountForms/VerifyForm.tsx
@@ -1,19 +1,17 @@
 import React, {useRef, useState} from 'react';
 import TextInput from '../../../components/formFields/TextInput';
 import {connect} from 'react-redux';
-import {
-    OAuth2Token,
-    Account,
-    ReduxState,
-    Message,
-} from '../../../utilities/types';
+import stateTypes from '../../../utilities/types/stateTypes';
 import {authPostRequest} from '../../../utilities/axiosClients';
 import dispatcher from '../../../utilities/dispatcher';
 import ButtonLink from '../../../components/links/ButtonLink';
 
 interface VerifyFormProps {
-    logIn(oauth2_token: OAuth2Token, account: Account): void;
-    openMessage(message: Message): void;
+    logIn(
+        oauth2_token: stateTypes.OAuth2Token,
+        account: stateTypes.Account,
+    ): void;
+    openMessage(message: stateTypes.Message): void;
     closeAllMessages(): void;
 }
 
@@ -118,7 +116,7 @@ function VerifyForm(props: VerifyFormProps) {
     );
 }
 
-const mapStateToProps = (state: ReduxState) => ({});
+const mapStateToProps = (state: stateTypes.ReduxState) => ({});
 const mapDispatchToProps = (dispatch: any) => ({
     logIn: dispatcher.logIn(dispatch),
     openMessage: dispatcher.openMessage(dispatch),
