+++ src/utilities/axiosClients.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import {SURVEY_BACKEND_URL, AUTH_BACKEND_URL} from '../constants';
-import {OAuth2Token} from './types';
+import stateTypes from './types/stateTypes';
 
 export function authPostRequest(url: string, data: {[key: string]: any}) {
     let formData = new FormData();
@@ -10,7 +10,10 @@ export function authPostRequest(url: string, data: {[key: string]: any}) {
     return axios.post(AUTH_BACKEND_URL + url, formData);
 }
 
-export function surveyGetRequest(url: string, oauth2_token: OAuth2Token) {
+export function surveyGetRequest(
+    url: string,
+    oauth2_token: stateTypes.OAuth2Token,
+) {
     return axios.get(SURVEY_BACKEND_URL + url, {
         headers: {
             Authorization: 'bearer ' + oauth2_token.access_token,
